refactor(cli): give gglib proxy its own handler module
Every other top-level command dispatches to a handlers::* module — serve
got one when it was rewritten for #632 — but proxy kept ~60 lines of
settings resolution and StandaloneProxyParams assembly inline in
dispatch.rs, the one match arm doing real work instead of delegating.

Extracts it to handlers/inference/proxy.rs, mirroring serve.rs: the two
are pinned and unpinned modes of the identical start_proxy_standalone
entry point, so reading them side by side now actually shows that. No
behavior change — subcommand routing (dashboard, cache-clear) stays in
dispatch.rs, matching how Chat routes its own History subcommand.
