+++ crates/gglib-cli/src/handlers/inference/README.md
@@ -7,9 +7,11 @@
 
 Inference command handlers.
 
-Handles `serve`, `chat`, and `question` — the three top-level commands
-that run models. Shared inference-config resolution and logging live
-in the [`shared`] submodule to avoid duplication.
+Handles `serve`, `proxy`, `chat`, and `question` — the top-level commands
+that run models. `serve` and `proxy` are the pinned and unpinned modes of
+the same `start_proxy_standalone` entry point, so their handlers mirror
+each other. Shared inference-config resolution and logging live in the
+[`shared`] submodule to avoid duplication.
 
 <!-- module-docs:end -->
 
@@ -21,6 +23,7 @@ in the [`shared`] submodule to avoid duplication.
 |--------|-----|------------|----------|
 | [`agent_question.rs`](agent_question.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-agent_question-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-agent_question-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-agent_question-coverage.json) |
 | [`chat.rs`](chat.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-chat-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-chat-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-chat-coverage.json) |
+| [`proxy.rs`](proxy.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-proxy-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-proxy-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-proxy-coverage.json) |
 | [`serve.rs`](serve.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-serve-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-serve-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-serve-coverage.json) |
 | [`shared.rs`](shared.rs) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-shared-loc.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-shared-complexity.json) | ![](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/mmogr/gglib/badges/gglib-cli-inference-shared-coverage.json) |
 <!-- module-table:end -->
