+++ crates/gglib-cli/src/handlers/inference/mod.rs
@@ -1,5 +1,6 @@
 #![doc = include_str!("README.md")]
 pub mod agent_question;
 pub mod chat;
+pub mod proxy;
 pub mod serve;
 pub mod shared;
